Move comparators to new file
